Move missing auth function to OTP unit
