+++ NitrokeyManager.cc
@@ -20,6 +20,7 @@
  */
 
 #include "libnitrokey/NitrokeyManager.h"
+#include "NitrokeyManager.h"
 #include "NitrokeyManagerOTP.h"
 #include "NitrokeyManagerPWS.h"
 #include "libnitrokey/LibraryException.h"
@@ -40,19 +41,6 @@ namespace nitrokey{
     std::mutex mex_dev_com_manager;
 
 using nitrokey::misc::strcpyT;
-
-
-    // package type to auth, auth type [Authorize,UserAuthorize]
-    template <typename S, typename A, typename T>
-    void NitrokeyManager::authorize_packet(T &package, const char *admin_temporary_password, shared_ptr<Device> device){
-      if (!is_authorization_command_supported()){
-        LOG("Authorization command not supported, skipping", Loglevel::WARNING);
-      }
-        auto auth = get_payload<A>();
-        strcpyT(auth.temporary_password, admin_temporary_password);
-        auth.crc_to_authorize = S::CommandTransaction::getCRC(package);
-        A::CommandTransaction::run(device, auth);
-    }
 
     shared_ptr <NitrokeyManager> NitrokeyManager::_instance = nullptr;
